attestation/file: normalize artifact paths to forward slashes
RecordArtifacts keyed the artifacts map on filepath.Rel output and, for
matched directories, appended string(os.PathSeparator). On Windows that
yields backslash-separated subject names (for example "subdir\file" and
"testdir\") while the same source tree on Linux or macOS produces forward
slashes. The recorded attestation subjects were therefore platform
dependent, which breaks cross-platform verification and policy subject
matching. The dirHashGlob matching had the same problem: globs use forward
slashes, so a pattern like "path/to/dir" never matched the backslash
relative path on Windows and directory hashing silently did not trigger.

Normalize relPath with filepath.ToSlash immediately after it is computed,
use a literal "/" for the directory-key suffix, and ToSlash the joined
symlink keys so every recorded key is forward-slashed on all platforms.
The product attestor already did this downstream (see attestation/product),
so this moves the normalization to where the keys are minted and also
covers the material attestor and the glob matching.

With the keys now platform independent, TestDirHash and TestDirHashWithSymlink
no longer fail on Windows, so move them out of the !windows build-tagged file
into file_test.go and drop the build tag. TestDirHashWithSymlink skips when
the OS does not permit symlink creation.

Fixes #590
